# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an auto-generated API Docs.",
        "title": "API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "your@mail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/token/renew": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Renew access and refresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "summary": "renew access and refresh tokens",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "refresh_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/user/sign/in": {
            "post": {
                "description": "Auth user and return access and refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "auth user and return access and refresh token",
                "parameters": [
                    {
                        "description": "Sign in payload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/user/sign/out": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "De-authorize user and delete refresh token from Redis.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "de-authorize user and delete refresh token from Redis",
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/user/sign/up": {
            "post": {
                "description": "Create a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "create a new user",
                "parameters": [
                    {
                        "description": "Sign up payload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    }
                }
            }
        },
        "/v1/wallet": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new crypto wallet with generated mnemonic and first blockchain address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Create a new wallet",
                "parameters": [
                    {
                        "description": "Create wallet payload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateWalletReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateWalletRes"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/core.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ApiResponse"
                        }
                    }
                }
            }
        },
        "/v1/wallet/restore": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Restore access to an existing wallet using secret phrase and optional passphrase.\nIf the wallet was protected with a passphrase, the correct passphrase must be provided.\nOn success, returns wallet identifier and associated blockchain addresses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Restore / Access existing wallet",
                "parameters": [
                    {
                        "description": "Restore wallet payload (secret phrase and optional passphrase)",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RestoreWalletReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet restored successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RestoreWalletRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid secret phrase or passphrase",
                        "schema": {
                            "$ref": "#/definitions/core.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.CreateWalletReq": {
            "type": "object",
            "required": [
                "wallet_name"
            ],
            "properties": {
                "passphrase": {
                    "type": "string"
                },
                "wallet_name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "dto.CreateWalletRes": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "secretPhrase": {
                    "type": "string"
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        },
        "dto.RestoreWalletReq": {
            "type": "object",
            "required": [
                "secret_phrase"
            ],
            "properties": {
                "passphrase": {
                    "type": "string"
                },
                "secret_phrase": {
                    "type": "string"
                },
                "wallet_name": {
                    "type": "string"
                }
            }
        },
        "dto.RestoreWalletRes": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        },
        "models.BlockchainAddress": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "addressId": {
                    "type": "string"
                },
                "createDate": {
                    "type": "string"
                },
                "updateDate": {
                    "type": "string"
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
        "models.SignIn": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "models.SignUp": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "user_role"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 255
                },
                "user_role": {
                    "type": "string",
                    "maxLength": 25
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "fromAddress": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "toAddress": {
                    "type": "string"
                },
                "transactionDate": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "string"
                },
                "wallet": {
                    "description": "🔗 Relation",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Wallet"
                        }
                    ]
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
        "models.Users": {
            "type": "object",
            "properties": {
                "createDate": {
                    "type": "string"
                },
                "createdUserId": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passwordHash": {
                    "type": "string"
                },
                "updateDate": {
                    "type": "string"
                },
                "updatedUserId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "userRole": {
                    "type": "string"
                },
                "userStatus": {
                    "type": "integer"
                },
                "wallets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Wallet"
                    }
                }
            }
        },
        "models.Wallet": {
            "type": "object",
            "properties": {
                "blockchainAddresses": {
                    "description": "🔗 Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BlockchainAddress"
                    }
                },
                "createDate": {
                    "type": "string"
                },
                "passphraseHash": {
                    "type": "string"
                },
                "secretPhraseHash": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                },
                "updateDate": {
                    "type": "string"
                },
                "walletId": {
                    "type": "string"
                },
                "walletName": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}